feat(csraw2): schema 2.4.0 — float32 angles for duel-engine pre-shot correction
The four view-angle fields fed into the duel engine's pre-shot
correction calc — weapon_fires.yaw_deg/pitch_deg and
first_sights.observer_yaw_deg/observer_pitch_deg — switch from int16
quantised at 1/100° to float32. Rounding to 1/100° lets the angular
delta drift up to ~0.014°, enough to flip individual shots across the
hard 2° threshold of PctCorrectionUnder2Deg for players whose median
pre-shot correction sits near the boundary. Measured on the 45-demo
personal corpus, this caused ~5pp divergence on PctCorrectionUnder2Deg
for one player on demo 14.

Mirrors the schema 2.3.0 position carve-out: any field fed into a
hard-thresholded duel-engine metric is float32; everything else stays
int16-quantised since median-based metrics tolerate sub-1/100° jitter.
The remaining first_sights.angle_deg / pitch_deg / yaw_deg deviation
fields stay int16.

Bridge gains yaw0to360F32 to normalise float32 yaws into the [0, 360)
range v1 structs expect.

Verified byte-exact on demo 14: PctCorrectionUnder2Deg 44.4/44.4 (was
44.4/38.9), MedianCorrectionDeg 2.232/2.232. Full corpus re-sweep
clean.

+++ internal/csraw2/types.go
@@ -83,7 +83,12 @@ type Round struct {
 // fed into the duel engine's distance computation — Damage.VictimPos and
 // WeaponFire.Pos — are float32 instead so that duels at exact distance-bin
 // boundaries (5/10/15/20/30 m) classify identically to v1's raw-float path.
-// View angles are int16 quantised to 1/100°. Velocities are int16 in u/s.
+// View angles are int16 quantised to 1/100° except for the four fields fed
+// into the duel engine's pre-shot correction calc — WeaponFire.YawDeg/
+// PitchDeg and FirstSight.ObserverYawDeg/ObserverPitchDeg — which are
+// float32 so PctCorrectionUnder2Deg classifies identically to v1 at the
+// 2° boundary. (Schema 2.4.0)
+// Velocities are int16 in u/s.
 
 // Kill is one row per kill event.
 type Kill struct {
@@ -148,9 +153,9 @@ type WeaponFire struct {
 	PosX          float32 `parquet:"pos_x"`
 	PosY          float32 `parquet:"pos_y"`
 	PosZ          float32 `parquet:"pos_z"`
-	YawDeg        int16 `parquet:"yaw_deg"`
-	PitchDeg      int16 `parquet:"pitch_deg"`
-	VelX          int16 `parquet:"vel_x"`
+	YawDeg        float32 `parquet:"yaw_deg"`
+	PitchDeg      float32 `parquet:"pitch_deg"`
+	VelX          int16   `parquet:"vel_x"`
 	VelY          int16 `parquet:"vel_y"`
 	VelZ          int16 `parquet:"vel_z"`
 	IsScoped      bool  `parquet:"is_scoped"`
@@ -183,8 +188,8 @@ type FirstSight struct {
 	AngleDeg          int16 `parquet:"angle_deg"`
 	PitchDeg          int16 `parquet:"pitch_deg"`
 	YawDeg            int16 `parquet:"yaw_deg"`
-	ObserverPitchDeg  int16 `parquet:"observer_pitch_deg"`
-	ObserverYawDeg    int16 `parquet:"observer_yaw_deg"`
+	ObserverPitchDeg  float32 `parquet:"observer_pitch_deg"`
+	ObserverYawDeg    float32 `parquet:"observer_yaw_deg"`
 	ObserverPosX      int16 `parquet:"observer_pos_x"`
 	ObserverPosY      int16 `parquet:"observer_pos_y"`
 	ObserverPosZ      int16 `parquet:"observer_pos_z"`
